refactor: improve handlers.rs modularity and readability
## Overview

Refactored the monolithic handlers.rs file (1,186 lines) into a modular structure with 9 focused modules, significantly improving code organization and maintainability.

## Major Changes

- 🏗️ **Refactor**: Split handlers.rs into functional modules
- ✨ **Feature**: Created modular handler architecture
- 🔧 **Improvement**: Enhanced code readability and maintainability
- 🏗️ **Refactor**: Extracted common traits and utilities

## Technical Details

- **Before**: Single file with 1,186 lines containing all handlers
- **After**: 9 focused modules with clear responsibilities:
  - : Alarm-related handlers (3 handlers, 137 lines)
  - : Variable operation handlers (5 handlers, 202 lines)
  - : Position-related handlers (5 handlers, 177 lines)
  - : System information handlers (8 handlers, 221 lines)
  - : I/O and register handlers (2 handlers, 78 lines)
  - : Job and movement handlers (5 handlers, 123 lines)
  - : File operation handlers (1 handler, 123 lines)
  - : Command handler registry (189 lines)
  - : Main module interface (15 lines)

- **Architecture**: Maintained existing CommandHandler trait and registry pattern
- **Performance**: No performance impact, same functionality with better organization

## Testing

- ✅ All existing tests pass (5/5 integration tests)
- ✅ Compilation successful with no warnings
- ✅ Clippy analysis clean
- ✅ Code formatting compliant
- ✅ Security audit passed
- ✅ No breaking changes to public API

## Breaking Changes

None - All existing functionality preserved with identical public interface.

## Related Issues

Improves code maintainability and developer experience for future enhancements.
