+++ moto-hses-mock/src/handlers/io.rs
@@ -1,0 +1,77 @@
+//! I/O and register command handlers
+
+use super::CommandHandler;
+use crate::state::MockState;
+use moto_hses_proto as proto;
+
+/// Handler for I/O operations (0x78)
+pub struct IoHandler;
+
+impl CommandHandler for IoHandler {
+    fn handle(
+        &self,
+        message: &proto::HsesRequestMessage,
+        state: &mut MockState,
+    ) -> Result<Vec<u8>, proto::ProtocolError> {
+        let io_number = message.sub_header.instance;
+        let service = message.sub_header.service;
+
+        match service {
+            0x0e => {
+                // Read
+                let value = if state.get_io_state(io_number) { 1 } else { 0 };
+                Ok(vec![value, 0, 0, 0])
+            }
+            0x10 => {
+                // Write
+                if message.payload.len() >= 4 {
+                    let value = i32::from_le_bytes([
+                        message.payload[0],
+                        message.payload[1],
+                        message.payload[2],
+                        message.payload[3],
+                    ]);
+                    state.set_io_state(io_number, value != 0);
+                }
+                Ok(vec![])
+            }
+            _ => Err(proto::ProtocolError::InvalidService),
+        }
+    }
+}
+
+/// Handler for register operations (0x79)
+pub struct RegisterHandler;
+
+impl CommandHandler for RegisterHandler {
+    fn handle(
+        &self,
+        message: &proto::HsesRequestMessage,
+        state: &mut MockState,
+    ) -> Result<Vec<u8>, proto::ProtocolError> {
+        let reg_number = message.sub_header.instance;
+        let service = message.sub_header.service;
+
+        match service {
+            0x0e => {
+                // Read
+                let value = state.get_register(reg_number);
+                Ok(value.to_le_bytes().to_vec())
+            }
+            0x10 => {
+                // Write
+                if message.payload.len() >= 4 {
+                    let value = i32::from_le_bytes([
+                        message.payload[0],
+                        message.payload[1],
+                        message.payload[2],
+                        message.payload[3],
+                    ]);
+                    state.set_register(reg_number, value);
+                }
+                Ok(vec![])
+            }
+            _ => Err(proto::ProtocolError::InvalidService),
+        }
+    }
+}
